Add possibility of having project files in different locations than root Add possibility of copying subdirectories into our project
ex [package.json]:
"react-tv": {
    "files": [
      "./public/index.html",
      "./dist/bundle.js",
      "./dist/assets/"
    ]
  }

./public/index.html file will be copied to [webos_project_root]/index.html
./dist/bundle.js file will be copied to [webos_project_root]/bundle.js
./dist/assets/ dir and it's content will be copied to [webos_project_root]/assets

note 1) in this change, .npmignore was not implemented

note 2) If you add a directory that stand on the root, it will not work. ex:

"react-tv": {
    "files": [
      "./public/index.html",
      "./dist/bundle.js",
      "./dist/assets/",
      "./videos/"
    ]
  }

This was not implemented cause may cause corruption of our default webos_project files, like appinfo.json or even the [].ipk file itself. Problem it could be implemented when the .npmignore.